feat: distance 타입 변경 Long -> Int

+++ src/main/kotlin/com/example/running/domain/running/controller/dto/EndRunningRequest.kt
@@ -1,7 +1,7 @@
 package com.example.running.domain.running.controller.dto
 
 class EndRunningRequest(
-    val distance: Long,
+    val distance: Int,
     val durationSec: Long,
     val cadence: Short,
     val heartRate: Short,
